+++ src/components/CalendarDateBox.tsx
@@ -1,0 +1,36 @@
+import { Notifications } from '@mui/icons-material';
+import { Box, Stack, Typography } from '@mui/material';
+
+import { Event } from '../types';
+
+export default function CalendarDateBox({
+  event,
+  isNotified,
+}: {
+  event: Event;
+  isNotified: boolean;
+}) {
+  return (
+    <Box
+      key={event.id}
+      sx={{
+        p: 0.5,
+        my: 0.5,
+        backgroundColor: isNotified ? '#ffebee' : '#f5f5f5',
+        borderRadius: 1,
+        fontWeight: isNotified ? 'bold' : 'normal',
+        color: isNotified ? '#d32f2f' : 'inherit',
+        minHeight: '18px',
+        width: '100%',
+        overflow: 'hidden',
+      }}
+    >
+      <Stack direction="row" spacing={1} alignItems="center">
+        {isNotified && <Notifications fontSize="small" />}
+        <Typography variant="caption" noWrap sx={{ fontSize: '0.75rem', lineHeight: 1.2 }}>
+          {event.title}
+        </Typography>
+      </Stack>
+    </Box>
+  );
+}
