+++ src/components/WeekView.tsx
@@ -1,0 +1,73 @@
+import {
+  Stack,
+  Table,
+  TableBody,
+  TableCell,
+  TableContainer,
+  TableHead,
+  TableRow,
+  Typography,
+} from '@mui/material';
+
+import { weekDays } from '../constants';
+import { Event } from '../types';
+import CalendarDateBox from './CalendarDateBox';
+import { formatWeek, getWeekDates } from '../utils/dateUtils';
+
+interface WeekViewProps {
+  currentDate: Date;
+  filteredEvents: Event[];
+  notifiedEvents: string[];
+}
+
+export default function WeekView({ currentDate, filteredEvents, notifiedEvents }: WeekViewProps) {
+  const weekDates = getWeekDates(currentDate);
+
+  return (
+    <Stack data-testid="week-view" spacing={4} sx={{ width: '100%' }}>
+      <Typography variant="h5">{formatWeek(currentDate)}</Typography>
+      <TableContainer>
+        <Table sx={{ tableLayout: 'fixed', width: '100%' }}>
+          <TableHead>
+            <TableRow>
+              {weekDays.map((day) => (
+                <TableCell key={day} sx={{ width: '14.28%', padding: 1, textAlign: 'center' }}>
+                  {day}
+                </TableCell>
+              ))}
+            </TableRow>
+          </TableHead>
+          <TableBody>
+            <TableRow>
+              {weekDates.map((date) => (
+                <TableCell
+                  key={date.toISOString()}
+                  sx={{
+                    height: '120px',
+                    verticalAlign: 'top',
+                    width: '14.28%',
+                    padding: 1,
+                    border: '1px solid #e0e0e0',
+                    overflow: 'hidden',
+                  }}
+                >
+                  <Typography variant="body2" fontWeight="bold">
+                    {date.getDate()}
+                  </Typography>
+                  {filteredEvents
+                    .filter((event) => new Date(event.date).toDateString() === date.toDateString())
+                    .map((event) => {
+                      const isNotified = notifiedEvents.includes(event.id);
+                      return (
+                        <CalendarDateBox key={event.id} event={event} isNotified={isNotified} />
+                      );
+                    })}
+                </TableCell>
+              ))}
+            </TableRow>
+          </TableBody>
+        </Table>
+      </TableContainer>
+    </Stack>
+  );
+}
